// src/common/open-api/models/UpgradeWorkflowRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpgradeWorkflowRequest = {
  name: string;
  taskOutput?: Record<string, Record<string, any>>;
  version?: number;
  workflowInput?: Record<string, Record<string, any>>;
};