// src/common/open-api/models/WorkflowStateUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { TaskResult } from './TaskResult';
export type WorkflowStateUpdate = {
  taskReferenceName?: string;
  taskResult?: TaskResult;
  variables?: Record<string, Record<string, any>>;
};